song_analysis/requirements.txt
librosa
soundfile
numpy
scipy
scikit-learn
umap-learn
hdbscan
matplotlib
pandas
tqdm